Add and implement Contexts
Contexts can be entered and left. When doing this, a normal Context implementation will notify an LksBhmGame's ContextManager. This ContextManager is a central part of the game that can be accessed to add ContextListeners. A ContextListener in turn will then be called whenever the Context that the listener is intended for (specified via a Context's class name) is being entered or exited. The Listener will also have access to the Context this way, so there are almost no restrictions what the Listener can do (it is restricted by the Context's api, though).

Contexts are supposed to power Tutorials later on, for example. This will work by configuring ContextListeners (Java code) inside Level description files (probably JSON), that wait for levels with specific IDs to be entered.