Add PuzzleWonListeners
By adding the new listener type, only one call to Puzzle.isSolved will be necessary in the future for all parties who are interested in such an event. This also should cut down on the number of PuzzleChangedListeners.
This commit also implements time- instead of event-based invalid tile marker updates. It also makes RandomPuzzleScreen dispose of its puzzle after it is won and the screen hidden instead of immediately, which makes it possible to set its puzzle to null, which is a cleaner state (also, of course, no disposed puzzle is drawn).