Implement cancelOnTap transition, although introducing a bug.
During implementation, it turned out that TransitionableScreens need a getInputProcessor() method, so that the input processor can be restored by a transition (that may be listeneing for input on its own) even when the screen's stage is not the (only) input processor of the screen, e.g. when there is a multiplexer for the backbutton handling.